server/requirements.txt
Flask
flask-cors
python-dotenv
PyPDF2
requests
requests
requests-oauthlib